requirements.txt
asgiref==3.8.1
Django==5.1.7
django-extensions==3.2.3
django-jazzmin==3.0.1
django-jet-reboot==1.3.10
feedparser==6.0.11
google-api-python-client==1.4.1
httplib2==0.22.0
MarkupSafe==3.0.2
oauth2client==4.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
rsa==4.9
sgmllib3k==1.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
Werkzeug==3.1.3
